[avalanche] Improve lock annotations for AvalancheFinalizeBlock
Summary:
The migration of the stake contender cache to the peer manager changed the locking expectations, so let's update the annotations to reflect this.
I didn't add the negative annotation for cs_main because it bubbles up everywhere which is more annoying than helpful, but kept the runtime assertion so debug build would catch it anyway.

Test Plan:
With clang:
  ninja all check

Reviewers: #bitcoin_abc, roqqit

Reviewed By: #bitcoin_abc, roqqit

Differential Revision: https://reviews.bitcoinabc.org/D17836