[electrum] fix opening Trezor settings menu while the device is locked
Summary:
For some reason this feature is `None` when the device is locked, while all other features seem to be properly defined.
Don't raise an error dialog, just  disable the passaphrase feature and add a tooltip to explain why it is disabled.

The error was:
```
    passphrases_label.setText(disen[features.passphrase_protection])
                              ~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
TypeError: list indices must be integers or slices, not NoneType
```

Test Plan: Open a Trezor wallet, lock the device, open the trezor setting menu (trezor logo in the bottom-right bar)

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17845