[cashtab] CSS tweak for default desktop agora view
Summary:
It was looking a little odd that the default desktop view for agora has 4 cards but they dont
take the full width.

Adjusting it to a 4 column grid when all the offers are not loaded.

We could just leave it always 4 columns, but 5 is nice on large screens when there is a lot of content

Test Plan: npm start, go to agora check the columns, loadallzeoffers and check the columns

Reviewers: #bitcoin_abc, emack, bytesofman

Reviewed By: #bitcoin_abc, emack, bytesofman

Subscribers: bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D17941