Work around a bug in clang < 13
Summary:
Our OSX CI currently uses clang 12 which is failing due to a bug. Fortunately it's easy to work around this bug and we already have an occurrence of this in our codebase.

The best solution would be to move to clang >= 13 as we aim to use C++20 in the future anyway, but for now this fixes the immediate issue.

Test Plan:
  ninja all check

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17977