[CI] Properly integrate the mocha/jest Junit reports with CI
Summary:
There is a naming convention for the Junit reports (junit_results*.xml) so use that so the CI will parse the tests failures and make it easier to spot the errors.

Also prevent the build from failing if the junit report is not available for any reason.

Test Plan:
  ./contrib/teamcity/build-configurations.py cashtab-tests

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D17990