[test] Finalize the tip when possible in abc_p2p_avalanche_contender_voting
Summary:
This will avoid having the polling slots filled with blocks, making the test slower and flaky.
Because this adds some rounds of vote we need to increase the stale factor to avoid stalling contenders prematurely.

This is a step towards fixing the test flakiness.

Depends on D18038.

Test Plan:
  ./test/functional/test_runner.py abc_p2p_avalanche_contender_voting

Reviewers: #bitcoin_abc, roqqit

Reviewed By: #bitcoin_abc, roqqit

Subscribers: roqqit

Differential Revision: https://reviews.bitcoinabc.org/D18039