[chronik] Electrum: handle unconfirmed parents in history
Summary:
This diff makes the chronik electrum history strictly compliant with the Electrum protocol, and also identical with Fulcrum ordering (because the protocol itself is not deterministic).
If this is found to be a problem in the future this can be revisited but for now we want to be a 1:1 replacement.

Depends on D18106 and D18107.

Test Plan:
  ./test/functional/test_runner.py chronik_electrum_*

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18108