[electrum] Remove support for the custom firmware on Trezor One
Summary:
The One is known to have compatibility issues on some devices with the latest trezorlib version. Deploying a custom firmware for this device that is not sold anymore and use a separate codebase will be a maintenance burden and is not worth the cost.
Also this device is old enough that it is very difficult to properly test it with all the versions that are available in the wild, so it's likely to generate a lot support requests.
The firmware remains available if people want to build it but this diff removes it from the Electrum distribution so that it is not officially supported.

Test Plan: Read the doc and check the file is no longer shipped.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18115