[chronik] Electrum: add the server.features endpoint
Summary:
According to the protocol, this should return the public URLs and a list of the available protocols/ports in order to remain connected from other peers.
For now a single URL is supported, via a command line option as the node has no way to determine this value on its own.
In order to avoid querying the genesis hash for every call, the value is cached and a bridge function is added to get the hash from the node where it is hardcoded.

Test Plan:
  ./test/functional/test_runner.py chronik_electrum_basic

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18123