[bip21 ecash spec] Tolerate unsupported params when token_id is specified
Summary:
For now, Cashtab is the only known user of the token_id param and this spec. There are a number of complications to token support in bip21. Token txs require specific OP_RETURN outputs, but these outputs depend on the user's utxo set, so wallets must construct them.

At the moment the spec is as simple as possible with the idea that it may be extended in the future. In all likelihood, it will change as wallet construction of token txs evolves to handle multiple protocols.

Test Plan: Proofread

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18124