[chronik] Bump Rust requirement to >= 1.87.0
Summary:
This will allow for using recent versions of some packages that we need as our dependencies for Chronik.

This implies fixing the new clippy issues that comes with the new version.
Most of them are trivial (comment spacing, replacing a calculation with an existing function, using type aliases).
The most involved part is to wrap large enums into a Box to avoid passing copies that consume lots of memory.

Also includes the Bcrypt library on windows as per D16547, thanks to @tobias_ruck.

Test Plan: Build everything that uses Rust with 1.87, run the GUIX builds.

Reviewers: #bitcoin_abc, PiRK, tobias_ruck

Reviewed By: #bitcoin_abc, PiRK, tobias_ruck

Subscribers: tobias_ruck

Differential Revision: https://reviews.bitcoinabc.org/D18135