[Cashtab] Show a splash screen while ChronikClient loads
Summary:
ChronikClient loads async now as we are getting the fastest server for each user. This is not handled in the App, so we get what looks like a "white screen of death" for ~1s while the process completes.

Handle with a splash screen.

We could (potentially) eliminate this altogether, e.g. by storing the best server in app storage, always loading with that one first, then checking in the background. But, because the app already takes a couple of seconds to unlock the screen (while latest utxo set loads), this is an acceptable solution. Even if we do not show a splash screen, the spinner time on the other screen would be that much longer.

Test Plan:
`npm test`, this is deployed at cashtab.io

verify we see the splashscreen instead of a white screen

imo testing this component directly would be overkill, it's about the change in user experience.

splashscreen (replacing blank white screen):

{F16209552}

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18188