[Cashtab] Add an info modal for airdrop msgs
Summary:
I do not think we can fully police "attackers" from sending scam messages to token holders using the airdrop feature. Anyone can send any message to any eCash users. So, users could also spoof airdrop msgs to people who do not even hold a token.

We should make it clear, though, that such messages are not necessarily official, and certainly not from ABC.

Add an info modal. Clicking the modal offers a scam warning and short description of what an airdrop message is.

Cashtab used to have "beware of scams!" on every single cashtab msg. It also used to not render msgs at all unless the user clicked through this warning. We could optionally bring something like this back, but I do not think it was effective. I am generally skeptical of doing anything that sounds helpful but is impossible to test for effectiveness, especially if it adds complication. If there is a scam msg on every msg, it is probably just as effective as no scam msg on any msg. The notice becomes ignored as noise.

There are probably more improvements available to this problem. But the info modal is a start.

Test Plan:
`npm test`, see screenshot, this diff is deployed at cashtab.io

New info button after "Airdrop Msg" text:

{F16332548}

on click:
{F16332557}

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18194